requirements.txt
streamlit==1.28.1
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.0.352
langchain-community==0.0.13
qdrant-client==1.7.0
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0
groq==0.4.1
streamlit-chat==0.1.1
webdriver-manager==4.0.1
lxml==4.9.3
pydantic==2.5.2
typing-extensions==4.8.0
huggingface-hub==0.19.4
